requirements.txt
streamlit
pyttsx3
gTTS
requests 
pytest
python-dotenv
